requirements.txt
datasets==3.0.2
langchain==0.3.4
peft==0.13.2
python_docx==1.1.2
torch==2.5.0+cu124
transformers==4.46.0
protobuf==3.20.3
sentencepiece==0.2.0
nltk==3.9.1